requirements.txt
django==1.9.2
Pillow
python-social-auth
django-filter
django-taggit
django-taggit-templatetags2